--- a/docs/stories/epic-10/story-10.5-position-management-system.md
+++ b/docs/stories/epic-10/story-10.5-position-management-system.md
@@ -1,0 +1,125 @@
+# Story 10.5: Position Management System
+
+**Epic 10: Core Trading Infrastructure - MVP Implementation**
+
+## Story Overview
+Implement a high-performance position management system that tracks all trading positions with real-time P&L calculation, margin monitoring, and position lifecycle management from open to close.
+
+## Acceptance Criteria
+
+### AC1: Real-Time Position Tracking
+- GIVEN active trading positions
+- WHEN market prices update
+- THEN system calculates real-time unrealized P&L for all positions
+- AND updates position metrics within 1 second of price changes
+- AND maintains position state consistency across all components
+
+### AC2: Position Opening
+- GIVEN an order fill event
+- WHEN creating or updating a position
+- THEN system handles new position creation and position pyramiding
+- AND calculates weighted average entry price for multiple entries
+- AND updates position size and margin requirements
+- AND completes position opening in <50ms
+
+### AC3: Position Closing
+- GIVEN a position close request
+- WHEN executing position closure
+- THEN system supports both partial and full position closes
+- AND calculates realized P&L accurately
+- AND completes position close in <100ms (95th percentile)
+- AND maintains audit trail of all position changes
+
+### AC4: Multi-Account Support
+- GIVEN multiple trading accounts
+- WHEN managing positions
+- THEN system maintains separate position tracking per account
+- AND provides consolidated portfolio views
+- AND supports account-specific margin calculations
+- AND handles cross-account risk aggregation
+
+### AC5: Position Aggregation
+- GIVEN multiple orders for the same instrument
+- WHEN orders are filled
+- THEN system aggregates positions by instrument per account
+- AND maintains accurate average entry price calculations
+- AND handles position reversals (long to short transitions)
+- AND tracks all contributing order IDs
+
+### AC6: Margin Monitoring
+- GIVEN open positions
+- WHEN calculating margin requirements
+- THEN system tracks margin used per position
+- AND provides total margin utilization per account
+- AND calculates available margin for new positions
+- AND integrates with OANDA margin calculations
+
+### AC7: Performance Metrics
+- GIVEN position operations
+- WHEN executing position management tasks
+- THEN system provides performance metrics and statistics
+- AND tracks total positions, open positions, and P&L summaries
+- AND monitors position manager resource usage
+- AND reports position synchronization status
+
+### AC8: Position Synchronization
+- GIVEN OANDA position data
+- WHEN synchronizing local state
+- THEN system maintains consistency with broker positions
+- AND resolves discrepancies automatically
+- AND provides fallback mechanisms for API failures
+- AND logs all synchronization events
+
+## Technical Implementation
+
+### Core Components
+- **PositionManager** (`execution-engine/app/positions/position_manager.py`)
+  - Real-time position tracking with background price updates
+  - Position lifecycle management (open/modify/close)
+  - Multi-account portfolio management
+  - Margin calculation and monitoring
+
+### Key Features
+- Asynchronous position monitoring with configurable update intervals
+- Position aggregation and average price calculation algorithms
+- Real-time P&L calculation using current market prices
+- Position close with support for partial and full closures
+- Background synchronization with OANDA position data
+- Performance tracking and metrics collection
+- Account summary generation with position data
+
+### Performance Requirements
+- Position opening: <50ms
+- Position closing: <100ms (95th percentile)
+- Price update processing: <1 second
+- Memory usage: <200MB for position manager
+- Support for 100+ concurrent positions per account
+
+### Data Models
+- **Position**: Core position entity with units, prices, and P&L
+- **PositionCloseRequest**: Request structure for position closure
+- **AccountSummary**: Comprehensive account state with positions
+- **PositionSide**: Enumeration for LONG/SHORT position sides
+
+### Integration Points
+- **OANDA API**: Position data, margin information, current prices
+- **Order Manager**: Position updates from order fills
+- **Risk Manager**: Position data for risk calculations
+- **Metrics System**: Position performance monitoring
+
+## Validation Status
+✅ **IMPLEMENTED** - Position management system is fully operational with real-time tracking, P&L calculation, and comprehensive position lifecycle management.
+
+## Files Modified/Created
+- `execution-engine/app/positions/position_manager.py` - Core position management logic
+- `execution-engine/app/positions/__init__.py` - Position module initialization
+- `execution-engine/app/core/models.py` - Position-related data models
+
+## Dependencies
+- **Story 10.1**: Market data integration for real-time pricing
+- **Story 10.2**: Execution engine for order processing
+- **Story 10.4**: Risk management for position validation
+
+## Story Points: 8
+## Priority: High
+## Status: ✅ COMPLETED